+++ packages/terra-list/src/SelectableList.jsx
@@ -4,11 +4,6 @@ import 'terra-base/lib/baseStyles';
 import List from './List';
 import SelectableUtils from './SelectableUtils';
 
-const KEYCODES = {
-  ENTER: 13,
-  SPACE: 32,
-};
-
 const propTypes = {
   /**
    * The children list items passed to the component.
@@ -31,7 +26,7 @@ const propTypes = {
    */
   onChange: PropTypes.func,
   /**
-   * An array of the currectly selected indexes
+   * An array of the currectly selected indexes.
    */
   selectedIndexes: PropTypes.array,
 };
@@ -87,7 +82,7 @@ class SelectableList extends React.Component {
     const initialOnKeyDown = item.props.onKeyDown;
 
     return (event) => {
-      if (event.nativeEvent.keyCode === KEYCODES.ENTER || event.nativeEvent.keyCode === KEYCODES.SPACE) {
+      if (event.nativeEvent.keyCode === SelectableList.Utils.keyCodes.ENTER || event.nativeEvent.keyCode === SelectableList.Utils.keyCodes.SPACE) {
         // The default isSelectable attribute is either undefined or true, unless the consumer specifies the item isSelectable attribute as false.
         if (item.props.isSelectable !== false) {
           this.handleOnChange(event, index);