+++ packages/terra-list/CHANGELOG.md
@@ -7,8 +7,11 @@ Unreleased
 ### Removed
 * Remove stylelint-disable selector-class-pattern
 
+### Added
+* SelectableList component
+
 ### Changed
-* Break out SelectableList Subcomponent
+* Update SingleSelectList and MultiSelectList components to use SelectableList
 
 1.10.0 - (September 19, 2017)
 ------------------